crew_mypage,modify,delete 개발중

+++ src/main/java/com/meta/overwash/domain/CrewDTO.java
@@ -13,7 +13,7 @@ import lombok.ToString;
 @ToString
 public class CrewDTO {
 
-	private Integer crewId;
+	private Long crewId;
 	private UserDTO user;
 	private Date crewBirth;
 	private String crewContact;
